Add and rework tests
